Organises files into packages
