+++ src/main/java/yapper/datastorage/DataStorage.java
@@ -1,14 +1,16 @@
-package main.java;
+package yapper.datastorage;
 
 import java.io.File;
 import java.io.FileWriter;
 import java.io.FileReader;
 import java.io.IOException;
-import java.util.Scanner;
 import java.io.FileNotFoundException;
-import main.java.categories.Task;
+import java.util.Scanner;
 import java.util.ArrayList;
 import java.util.Collections;
+import yapper.taskTypes.Task;
+import yapper.taskTypes.TaskList;
+import yapper.parser.Parser;
 
 public class DataStorage {
     private File dataFile;
